Fix typo in offset parameter (#39)

+++ blobit-core/src/main/java/org/blobit/core/api/BucketHandle.java
@@ -157,7 +157,7 @@ public interface BucketHandle {
      */
     NamedObjectDownloadPromise downloadByName(String name,
             Consumer<Long> lengthCallback,
-            OutputStream output, int offset,
+            OutputStream output, long offset,
             long length);
 
     /**
