+++ blobit-core/src/main/java/org/blobit/core/mem/LocalManager.java
@@ -382,7 +382,7 @@ public class LocalManager implements ObjectManager {
         public NamedObjectDownloadPromise downloadByName(String name,
                                                          Consumer<Long> lengthCallback,
                                                          OutputStream output,
-                                                         int offset, long length) {
+                                                         long offset, long length) {
             List<String> ids = null;
             try {
                 ids = objectNames.get(name);
